refactor: types 폴더 삭제 & 경로 수정 & 잡다한 수정내용
