+++ daycan-front/apps/admin/src/components/index.ts
@@ -4,10 +4,13 @@ export * from "./Sidebar";
 export * from "./PageToolbar";
 export * from "./MenuItemHeader";
 export * from "./FilterSearchbar";
-export * from "./DropDownChip";
 export * from "./CareSheetDataSummaryCard";
 export * from "./ReportDataView";
 export * from "./CareSheetDataView";
 export * from "./Filter";
 export * from "./ListHeaderLayout";
 export * from "./ListItemLayout";
+export * from "./EditAuthModal";
+export * from "./DeleteConfirmModal";
+export * from "./GenderSelector";
+export * from "./CareLevelRoleDropDownSelector";
